Move Executor from settings, Fixed CeleryExecutor so it sends messages to workers, Added CLI cupport for scheduler
